Update Tester.py
When New is runned before the Open within a single run of the program, Open combines free space and the figure, decreasing the scaled size of 3d shape.

Long distances are less observable if the program is ended with an empty "RaspS.stl" or started with an empty "RaspS.stl". The second option can be implemented easier before starting MainWin().